LibC: Use 64-bit stack smash value for 64-bit mode
Otherwise it'll use the first 32 bits that happen to come after,
leading to very weird bugs. Fixes #8601

+++ Userland/Libraries/LibC/crt0.cpp
@@ -14,7 +14,7 @@
 #ifndef _DYNAMIC_LOADER
 extern "C" {
 
-extern u32 __stack_chk_guard;
+extern size_t __stack_chk_guard;
 
 int main(int, char**, char**);
 
@@ -31,7 +31,7 @@ NAKED void _start(int, char**, char**)
 
 int _entry(int argc, char** argv, char** env)
 {
-    u32 original_stack_chk = __stack_chk_guard;
+    size_t original_stack_chk = __stack_chk_guard;
     arc4random_buf(&__stack_chk_guard, sizeof(__stack_chk_guard));
 
     if (__stack_chk_guard == 0)
